fix(b4): pf in nfqnl_msg_config_cmd is __be16, was emitted as 1 byte
cmd.pf must be a 16-bit AF_* value; emitting u8 truncated the attribute to
3 bytes (len=7) so the kernel could not parse the PF_BIND command — the queue
was never registered. Cast to u16 big-endian.

+++ crates/balansir-b4/src/nfqueue.rs
@@ -133,7 +133,8 @@ impl NfQueue {
         let mut cmd_struct = Vec::with_capacity(4);
         cmd_struct.push(cmd);
         cmd_struct.push(0);
-        cmd_struct.extend_from_slice(&pf.unwrap_or(0).to_be_bytes());
+        // pf is __be16 (AF_*), not a single byte.
+        cmd_struct.extend_from_slice(&(pf.unwrap_or(0) as u16).to_be_bytes());
         push_nla(&mut attrs, NFQA_CFG_CMD, &cmd_struct);
 
         self.netlink_msg(NFQNL_MSG_CONFIG, attrs, res_id)
